*Carson* has received its `1.0` update!
# What's new?
### Ah jeez, what *hasn't* changed since release 0.4?

- The *Carson* star itself has been given tweaks and a fresh coat of paint.
- Added new nebular skybox.
- All planets have been completely rebuilt and revamped with new terrain as well as given proper generated heightmaps. No more taking the standard terrain texture and grayscaling it, save a few tweaks here and there.
-  Increased heightmap resolution to 1024 at the equator thanks to the introduction of LODs allowing for higher resolution while keeping frames stable. It allows for mountains, cracks, craters, and shorelines to be much more defined, and helps to keep things detailed when on the surface.
- Added another new planet named *Dester*.
- Added props such as trees and cacti to some planets, further helping to keep things detailed when on the surface.
- Added rain to Pyer.
- Tweaked gas giants.
- **New original flight music composed by yours truly (will likely post it on my YouTube music channel soon)!**

# What now?
### Well, a couple things...
The original idea behind *Carson* was that I wanted to create a star system that felt immersive.  The celestial bodies showcased in *Outer Wilds* as well as the beautiful star systems made by other talented people are rather tiny; you can walk across the circumference of a planet in about a minute give or take, which isn't big at all.

I figured the best thing to do was make a much bigger star system, but also without being *too* big. It's a very sensitive scale to balance, mind you, because if you make it too big then it'll just become boring and likely frustrating. But too small and it breaks immersion.

*Carson* was inspired by the *Kerbol* system (as well as modded solar systems via *Kopernicus*) as shown in the famous spacecraft simulator *Kerbal Space Program*, as seen by *Carson*'s overall planetary designs.

What I wanted to figure out was just *how big* to make my objects. I settled around the size they are now. I estimated it to be the sweet spot for size. Again, I wanted my system to be big enough to be immersive, but not too big to be frustrating for people or boring. I had to take into consideration that the ship, vanilla-wise anyways, is pretty slow too.

Obviously, this mod has its flaws; a lot of them, mind you. But I've had fun developing this mod and some of the wonderful people like Trifid on the *Outer Wilds Modding* Discord server appreciate the creative decisions I took when developing it. After all, if I'm going to make a planet, *I'm going to make a planet*. Because planets are huge.

I hope you've enjoyed the effort I've put into this regardless. It definitely does something different from other systems in *New Horizons*, and that was pretty much the whole scope of the mod.

### This is not the end.
I want to clarify that I am nowhere near finished with this mod. Plenty of updates are planned for the future. I just figured that the current version, with all the blood, sweat, and tears put into it, was a nice stable update to release with. More celestial bodies or maybe even more stars are planned. We'll see. But I also don't want my mod to get *too* bloated file-wise. Unfortunately, most systems made with *New Horizons* grow outdated and broken as support is dropped for them, whether because of personal reasons or whatever else. I'll be doing my best to keep my system up-to-date and fix anything that breaks in the future. I'll need to anyways if I'm going to add new content.

Thanks for downloading *Carson*, and stay tuned for more!